+++ examples/GetBlockItemStatus/Program.cs
@@ -12,8 +12,12 @@ internal sealed class GetBlockItemSummaryOptions
     [Option(HelpText = "Transaction hash to lookup", Required = true)]
     public string TransactionHash { get; set; }
 
-    [Option(HelpText = "URL representing the endpoint where the gRPC V2 API is served.", Required = true)]
+    [Option(HelpText = "URL representing the endpoint where the gRPC V2 API is served.", Required = true,
+        Default = "http://node.testnet.concordium.com/")]
     public string Endpoint { get; set; }
+
+    [Option(HelpText = "Port for the gRPC V2 API.", Required = true, Default = 20_000)]
+    public ushort Port { get; set; }
 }
 
 internal static class ExampleHelpers
@@ -43,7 +47,8 @@ public static class Program
     /// An example showing how one can query transaction status from a node.
     /// </summary>
     /// <param name="args">GetBlockItemSummaryOptions
-    /// Example: --endpoint http://node.testnet.concordium.com/ --transactionhash 143ca4183d0bb204000ad08e0fd5792985c808861b97f3b81cb9016ad39d09d2</param>
+    /// Example: --endpoint http://node.testnet.concordium.com/ --transactionhash 143ca4183d0bb204000ad08e0fd5792985c808861b97f3b81cb9016ad39d09d2 --port 20000
+    /// </param>
     public static async Task Main(string[] args)
     {
         if (!ExampleHelpers.TryParse(args, out GetBlockItemSummaryOptions? options))
@@ -53,11 +58,12 @@ public static class Program
 
         var endpoint = new Uri(options!.Endpoint);
         var transactionHash = TransactionHash.From(options.TransactionHash);
+        var port = options.Port;
 
-        using var client = new ConcordiumClient(endpoint, 20_000);
+        using var client = new ConcordiumClient(endpoint, port);
 
         Console.WriteLine("Query node...");
-        var transactionStatus = await client.GetBlockItemStatus(transactionHash);
+        var transactionStatus = await client.GetBlockItemStatusAsync(transactionHash);
 
         // Conditional on the transaction status different actions can be handled.
         switch (transactionStatus)
@@ -71,13 +77,15 @@ public static class Program
             case TransactionStatusCommitted committed:
                 HandleTransactionStatusCommitted(committed);
                 break;
-        };
+            default:
+                throw new NotImplementedException(transactionStatus.GetType().ToString());
+        }
     }
 
     /// <summary>
     /// Example how one can handle received state.
     /// </summary>
-    private static void HandleTransactionStatusReceived(TransactionStatusReceived transactionStatusReceived) => Console.WriteLine("Transaction received");
+    private static void HandleTransactionStatusReceived(TransactionStatusReceived _) => Console.WriteLine("Transaction received");
 
     /// <summary>
     /// Example how one can handle finalized state. One can use the BlockItemSummary to get additional information like